+++ src/OpenTelemetry.Shims.OpenTracing/TracerShim.cs
@@ -17,6 +17,7 @@
 using System;
 using System.Collections.Generic;
 using global::OpenTracing.Propagation;
+using OpenTelemetry.Context;
 using OpenTelemetry.Context.Propagation;
 
 namespace OpenTelemetry.Shims.OpenTracing
@@ -83,7 +84,12 @@ namespace OpenTelemetry.Shims.OpenTracing
                 propagationContext = this.textFormat.Extract(propagationContext, carrierMap, GetCarrierKeyValue);
             }
 
-            return !propagationContext.ActivityContext.IsValid() ? null : new SpanContextShim(new Trace.SpanContext(propagationContext.ActivityContext), propagationContext.ActivityBaggage);
+            // TODO:
+            //  Not sure what to do here. Really, Baggage should be returned and not set until this ISpanContext is turned into a live Span.
+            //  But that code doesn't seem to exist.
+            // Baggage.Current = propagationContext.Baggage;
+
+            return !propagationContext.ActivityContext.IsValid() ? null : new SpanContextShim(new Trace.SpanContext(propagationContext.ActivityContext));
         }
 
         /// <inheritdoc/>
@@ -115,7 +121,7 @@ namespace OpenTelemetry.Shims.OpenTracing
             if ((format == BuiltinFormats.TextMap || format == BuiltinFormats.HttpHeaders) && carrier is ITextMap textMapCarrier)
             {
                 this.textFormat.Inject(
-                    new PropagationContext(shim.SpanContext, shim.GetBaggageItems()),
+                    new PropagationContext(shim.SpanContext, Baggage.Current),
                     textMapCarrier,
                     (instrumentation, key, value) => instrumentation.Set(key, value));
             }